Docker/requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient 
sqlparse
pytz
asgiref
psycopg2-binary
